AppBar: Maintenance changes, Imageview changes
In the past a custom ImageView was used to force the image in the
collapsing layout to be squared. However, with recent themes the
collapsing layout is unwilling to be forced to exact dimensions and
always adds the amount of height of the status bar on top.

The change allows the image to be slightly cropped but enforces at least
the width of the image as height.

This commit also changes the use from androidx toolbar to Material Components
toolbar. This way all components involved are from the same package
"com.google.android.material".

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>