MPD: Change behavior for albumartist use
Before the change M.A.L.P. had the option to list only AlbumArtists in
the overview. But upon selection of an AlbumArtists all albums matching
the artist and albumartist are listed.

This commit changes the behavior, so that it only lists albums with the
albumartist matching the selected one (and albums with no albumartist
tag but matching artist tag, which is MPD behavior).

Addresses #289

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>