MPDFileFilter: Fix album track filtering for tracks with no ArtistSort
If the option to use ArtistSort tag is active albums are visible for
which no artistsort tag is set because of MPD's fallback mechanism.

However, the filtering from the M.A.L.P. does not take that into account
and when the file list is retrieved tracks are discarded because only
the "artist" tag is set.

This commit addresses the problem by falling back to artist tag if
neither "artistsort" or "albumartistsort" is set.

Fixes #285

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>