MPD/Library: Clean up artist, album artist sort logic
The logic to fetch, select, parse is distributed across the whole code
base which makes it a nightmare to maintain and fix bugs.

This commit streamlines the whole thing and moves the logic as far down
to the backend as possible so that the UI is not involved in the logic
besides selecting it based on the preference.

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>