MPD/State: Add proper state monitoring to MPD
Before that commit MALP was only able to handle all state changes with
the same callback. This commit changes it by introducing glue code that
enables differentiating between what changed in a state.

This unblocks fine granular tracking of server-side changes and is
currently used in the outputs and profiles fragments.

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>