Fragments: React to DATABASE and STORED_PLAYLIST changes
This commit wires up the BaseFragment to DATABASE notification which
means that all fragments will refresh their content if MPD updates the
database (e.g. new albums added).

Also the stored playlist fragment is monitoring for changes of
STORED_PLAYLISTS so if a playlist is added or removed it is visible.

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>