MPDConnection: Fix race connection leading to disconnect
Recent changes to the MPDConnection idling parsing made changes to the
timing of deidling.

The race happened between stopIdle() spawning the read time out task and
an idle request ending from the server side.

It is technically not possible to know when MPD deidles because of
external events.

A previous time out task was not cancelled although the MPDConnection
went into idle state again. When requested to deidle in the time frame
before the old time out task fired the stopIDLE function bailed out and
not sent the noidle command and then actually resulting in a timeout.

Two changes are done to prevent that from happening:

- Guard the whole deidling after actually reading the buffer with
  synchronize on the MPDConnection. This serializes the cancellation of
  the read timeout task of the stopIDLE task.
- Do not bail out if a time out task is already scheduled when deidling.
  This prevents "noidle" not being send even if there was a race before,
  successfully recovering the connection.

Both changes are not preventing that a "noidle" command is send if the
server-side deidle coincides with a user requested deidle. But MPD is
robust against that because it is technically not possible to prevent
that.

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>