MPD: Do not invalidate cache on disconnect
When going in and out of the application the MPD connection is
disconnected. This was negating large part of the cache and resulted in
memory usage of the application increasing faster than the garbage
collector being able to clean up because of the high number of String
objects that are allocated during protocol parsing.

Signed-off-by: Hendrik Borghorst <hendrikborghorst@gmail.com>